requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
Faker==30.3.0
ipython==8.28.0
pre_commit==4.0.1
django-bootstrap-v5==1.0.11
pytest-django==4.9.0
pytest-factoryboy==2.7.0
freezegun==1.5.1
parameterized==0.9.0
isort==5.13.2
pillow==10.4.0
celery==5.4.0
redis==5.1.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-celery-beat==2.7.0
djangorestframework==3.15.2
